Standardize toolbar for layout and editing menus
The zoom tool is refactored to a reusable component, toolbar buttons on both menus are at the top, and page navigation uses the same buttons.
Returning to the main menu is done exclusively through the "Voltar" button, with the path at the top disabled within menus, to ensure the "your changes will be lost" prompt always appears when the user wants to exit.
The altered components are more compartmentalized with more props and less bloating of state.

AdventurousGui/updates-OCR#11 AdventurousGui/updates-OCR#28 AdventurousGui/updates-OCR#1