parametrize large upload temp storage path; alter param name for upload progression
Using info["progress"] to indicate the progression of the upload can be confused with info["ocr"]["progress"]. Now the client is aware of the upload state with "stored", which rises from 0% to 100% as chunks are stored and becomes "true" when all the pages are extracted to individual files. A bug is fixed where the info for a large file (more than 1 chunk) would lack the "type" field.

AdventurousGui/updates-OCR#6