allow configure and request OCR of folder's content excl. subfolders
An OCR configuration can be created, that will be applied when requesting OCR of the folder.

OCR will be performed on the documents within the folder (SUBFOLDERS EXCLUDED) using the configuration setup for the folder (default or custom).

Each document retains its custom configuration in _data.json, but it is NOT USED, in favour of the folder's config.

AdventurousGui/updates-OCR#19
closes AdventurousGui/updates-OCR#31
some additions relate to AdventurousGui/updates-OCR#27