rewrite client side of file upload and allow higher chunk sizes
Checking for upload status was complex and not asynchronous, updating the status only after a certain number of chunks and causing "state.info is undefined" errors due to the removal of filesystem info from the server response. The upload status is now checked asynchronously every 5 seconds while the uploaded file is not fully stored, and no longer causes errors.

NGINX now generally limits client_max_body_size to a generous 2MB and reserves a higher limit for the two endpoints which can receive large user-uploaded files. For these, the limits have been increased in docker-compose files from 200MB to 2GB, as small chunk sizes simply increased network traffic. The browser client splits file into chunks of 1GB size. This now ensures the vast majority of use cases only require one or two requests to /upload-file, reducing the risk of users interrupting the upload process.

For simplification, and due to no evident difference in performance, updates to state.info no longer try to limit themselves to the rows corresponding to files being uploaded. All the data is simply updated more frequently while files are being uploaded.

Stuck uploads were also not registering due to being based on the document's "creation" field, which was set only after the end of preparation for OCR. Now, it is set on prepare-upload. The check for stuck uploads is performed on page load to avoid being dependent on a user keeping the website open for longer than the check period.

closes AdventurousGui/updates-OCR#40