hotfix: replace non-functional invisible font with invisible Times New Roman
The special invisible font had to be replaced, as all the text placed on the result PDFs appeared to be spaces due to an unknown past change. With this change, the text is available for search and is shown as Times New Roman with size 8 when copy-pasted to a rich text editor. The font is preloaded on system start to avoid lazy loading on first OCR.